pyproject.toml
[tool.poetry]
name = "project2"
version = "0.1.0"
description = ""
authors = ["Ramazan <syrlybai056@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
attrs = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
